feat: docs desc can support markdown files
